# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Indo-German Mobility Corridor Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employers/me:
    get:
      operationId: getMyEmployerProfile
      tags: [employers]
      summary: Get my employer profile
      responses:
        "200":
          description: Employer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employer"
    patch:
      operationId: updateMyEmployerProfile
      tags: [employers]
      summary: Update my employer profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employer"
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List my documents
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    patch:
      operationId: updateDocument
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}/verify:
    post:
      operationId: verifyDocument
      tags: [documents]
      summary: Verify or reject a document (admin/facilitator)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentVerification"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  # ── TRAINING ─────────────────────────────────────────────────────────────────
...
  /interviews:
    get:
      operationId: listInterviews
      tags: [interviews]
      summary: List interviews
      responses:
        "200":
          description: Interviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Interview"
    post:
      operationId: scheduleInterview
      tags: [interviews]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewInput"
      responses:
        "201":
          description: Scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
  /interviews/{id}:
    get:
      operationId: getInterview
      tags: [interviews]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Interview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
    patch:
      operationId: updateInterview
      tags: [interviews]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"

  # ── OFFERS ───────────────────────────────────────────────────────────────────
  /offers:
    get:
      operationId: listOffers
      tags: [offers]
      summary: List job offers
      responses:
        "200":
          description: Offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
    post:
      operationId: createOffer
      tags: [offers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /offers/{id}:
    get:
      operationId: getOffer
      tags: [offers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
    patch:
      operationId: updateOffer
      tags: [offers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"

  # ── WELFARE / GRIEVANCE ───────────────────────────────────────────────────────
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # ── ANALYTICS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Employer:
      type: object
      required: [id, userId, companyName, country, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        companyName:
          type: string
        country:
          type: string
        city:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        contactName:
          type: string
        contactEmail:
          type: string
        verificationStatus:
          type: string
          enum: [pending, verified, rejected]
        createdAt:
          type: string
    EmployerUpdate:
      type: object
      properties:
        companyName:
          type: string
        city:
          type: string
        industry:
          type: string
        contactName:
          type: string
...
    Document:
      type: object
      required: [id, candidateId, type, name, verificationStatus, uploadedAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        type:
          type: string
          enum: [passport, aadhaar, education_certificate, marksheet, iti_diploma, experience_letter, resume, language_certificate, police_clearance, medical_certificate, offer_letter, recognition_document, visa_document, insurance]
        name:
          type: string
        fileUrl:
          type: ["string", "null"]
        verificationStatus:
          type: string
          enum: [pending, under_review, verified, rejected, expired]
        expiryDate:
          type: ["string", "null"]
        reviewerNotes:
          type: ["string", "null"]
        uploadedAt:
          type: string
    DocumentInput:
      type: object
      required: [type, name]
      properties:
        type:
          type: string
        name:
          type: string
        fileUrl:
          type: string
        expiryDate:
          type: string
    DocumentUpdate:
      type: object
      properties:
        name:
          type: string
        fileUrl:
          type: string
        expiryDate:
          type: string
    DocumentVerification:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [verified, rejected]
        reviewerNotes:
          type: string

    # ── TRAINING ──
...
    Interview:
      type: object
      required: [id, applicationId, candidateId, employerId, scheduledAt, status, createdAt]
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        candidateId:
          type: integer
        candidateName:
          type: ["string", "null"]
        employerId:
          type: integer
        employerName:
          type: ["string", "null"]
        vacancyTitle:
          type: ["string", "null"]
        scheduledAt:
          type: string
        meetingLink:
          type: ["string", "null"]
        status:
          type: string
          enum: [scheduled, completed, cancelled, no_show]
        feedback:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
        createdAt:
          type: string
    InterviewInput:
      type: object
      required: [applicationId, scheduledAt]
      properties:
        applicationId:
          type: integer
        scheduledAt:
          type: string
        meetingLink:
          type: string
    InterviewUpdate:
      type: object
      properties:
        status:
          type: string
        feedback:
          type: string
        outcome:
          type: string
        meetingLink:
          type: string

    # ── OFFER ──
    Offer:
      type: object
      required: [id, applicationId, candidateId, employerId, status, createdAt]
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        candidateId:
          type: integer
        candidateName:
          type: ["string", "null"]
        employerId:
          type: integer
        employerName:
          type: ["string", "null"]
        vacancyTitle:
          type: ["string", "null"]
        salaryOffered:
          type: ["integer", "null"]
        joiningDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [issued, accepted, rejected, expired, withdrawn]
        offerLetterUrl:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
    OfferInput:
      type: object
      required: [applicationId, candidateId, employerId]
      properties:
        applicationId:
          type: integer
        candidateId:
          type: integer
        employerId:
          type: integer
        salaryOffered:
          type: integer
        joiningDate:
          type: string
        offerLetterUrl:
          type: string
    OfferUpdate:
      type: object
      properties:
        status:
          type: string
        joiningDate:
          type: string

    # ── WELFARE ──
...
    Notification:
      type: object
      required: [id, userId, type, title, read, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        title:
          type: string
        body:
          type: ["string", "null"]
        read:
          type: boolean
        link:
          type: ["string", "null"]
        createdAt:
          type: string

    # ── ANALYTICS ──
...